+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 ## Changed
 
 - 📝(doc) minor README.md formatting and wording enhancements
+- ♻️Stop setting a default title on doc creation #634
+
 
 ## [2.2.0] - 2025-02-10